Move functions to UtilFunctions.py.
