+++ src/Mod/Assembly/CommandInsertLink.py
@@ -30,34 +30,12 @@ if App.GuiUp:
     import FreeCADGui as Gui
     from PySide import QtCore, QtGui, QtWidgets
 
+import UtilsAssembly
 # translate = App.Qt.translate
 
 __title__ = "Assembly Command Insert Link"
 __author__ = "Ondsel"
 __url__ = "https://www.freecad.org"
-
-
-def activeAssembly():
-    doc = Gui.ActiveDocument
-
-    if doc is None or doc.ActiveView is None:
-        return None
-
-    active_part = doc.ActiveView.getActiveObject("part")
-
-    if active_part is not None and active_part.Type == "Assembly":
-        return active_part
-
-    return None
-
-
-def isDocTemporary(doc):
-    # Guard against older versions of FreeCad which don't have the Temporary attribute
-    try:
-        docTemporary = doc.Temporary
-    except AttributeError:
-        docTemporary = False
-    return docTemporary
 
 
 class CommandInsertLink:
@@ -80,10 +58,10 @@ class CommandInsertLink:
         }
 
     def IsActive(self):
-        return activeAssembly() is not None
+        return UtilsAssembly.activeAssembly() is not None
 
     def Activated(self):
-        assembly = activeAssembly()
+        assembly = UtilsAssembly.activeAssembly()
         if not assembly:
             return
         view = Gui.activeDocument().activeView()
@@ -139,7 +117,7 @@ class TaskAssemblyInsertLink(QtCore.QObject):
         docList = App.listDocuments().values()
 
         for doc in docList:
-            if isDocTemporary(doc):
+            if UtilsAssembly.isDocTemporary(doc):
                 continue
 
             for obj in doc.findObjects("App::Part"):
