+++ src/Mod/Assembly/UtilsAssembly.py
@@ -1,0 +1,56 @@
+# SPDX-License-Identifier: LGPL-2.1-or-later
+# /****************************************************************************
+#                                                                           *
+#    Copyright (c) 2023 Ondsel <development@ondsel.com>                     *
+#                                                                           *
+#    This file is part of FreeCAD.                                          *
+#                                                                           *
+#    FreeCAD is free software: you can redistribute it and/or modify it     *
+#    under the terms of the GNU Lesser General Public License as            *
+#    published by the Free Software Foundation, either version 2.1 of the   *
+#    License, or (at your option) any later version.                        *
+#                                                                           *
+#    FreeCAD is distributed in the hope that it will be useful, but         *
+#    WITHOUT ANY WARRANTY; without even the implied warranty of             *
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU       *
+#    Lesser General Public License for more details.                        *
+#                                                                           *
+#    You should have received a copy of the GNU Lesser General Public       *
+#    License along with FreeCAD. If not, see                                *
+#    <https://www.gnu.org/licenses/>.                                       *
+#                                                                           *
+# ***************************************************************************/
+
+import FreeCAD as App
+
+if App.GuiUp:
+    import FreeCADGui as Gui
+
+# translate = App.Qt.translate
+
+__title__ = "Assembly utilitary functions"
+__author__ = "Ondsel"
+__url__ = "https://www.freecad.org"
+
+
+def activeAssembly():
+    doc = Gui.ActiveDocument
+
+    if doc is None or doc.ActiveView is None:
+        return None
+
+    active_part = doc.ActiveView.getActiveObject("part")
+
+    if active_part is not None and active_part.Type == "Assembly":
+        return active_part
+
+    return None
+
+
+def isDocTemporary(doc):
+    # Guard against older versions of FreeCad which don't have the Temporary attribute
+    try:
+        temp = doc.Temporary
+    except AttributeError:
+        temp = False
+    return temp
